requirements.txt
amqp==5.3.1
ansi2html==1.9.2
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2025.4
cstruct==6.1
Django==5.2.1
django-crispy-forms==2.4
dnspython==2.7.0
docker==7.1.0
idna==3.10
inflate64==1.0.3
jefferson==0.4.6
kombu==5.5.4
lzallright==0.2.6
multivolumefile==0.2.3
packaging==25.0
prompt_toolkit==3.0.51
psutil==7.0.0
py7zr==1.0.0
pybcj==1.0.6
pycryptodomex==3.23.0
pymongo==4.13.0
pyppmd==1.2.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
PyYAML==6.0.2
pyzstd==0.17.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
texttable==1.7.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
